requirements.txt
Django==3.2.9
Pillow==8.4.0
gunicorn==20.1.0
whitenoise==5.3.0
sentry-sdk==1.4.3
requests==2.26.0
django-redis==5.0.0
pytz==2021.3
celery==5.2.0
celery-redbeat==2.0.0
psycopg2-binary==2.8.6
arrow==1.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
django-environ==0.8.1
django-storages==1.12.3
django-cors-headers==3.10.0
django-braces==1.15.0
django-extensions==3.1.5
django-post-office==3.5.3
django-crispy-forms==1.13.0
crispy-tailwind==0.5.0
django-allauth==0.45.0
boto3==1.20.3
boto3-stubs==1.20.3
django-import-export==2.6.1
honeybadger==0.8.0
django-ses==2.3.0
djangoql==0.15.4
flake8==4.0.1
django-ckeditor==6.1.0
django-taggit==1.5.1
django-filter==21.1
drf-spectacular==0.21.0
django-tailwind==2.2.2
django-configurations==2.3.1
dj-database-url==0.5.0
django-cache-url==3.2.3